Will acquire the alternative data streams for SDS and usnjrnl
These were skipped, due to an error introduced in the TarLoader.
The tarloader attempts to lstat the entry, where the ADS wasn't assigned.
These files did not have a ATTRIBUTE_TYPE_CODE.DATA attribute. E.g:
  - $Extend/$Usnjrnl
  - $Secure

This is why they returned an error.
Now, we get the entry directly, which does have the `ads` property set.
Fixing the issue.