Will acquire the alternative data streams for SDS and usnjrnl (#52)
In #33 there was a change to the tarloader that allows the collection of volatile data. This code uses lstat() to get some properties of the file that is collected, however it fails on ntfs files that only have alternative data streams and no main data stream (e.g. $Secure).

When explicitly getting the alternate data stream as a file entry (using e.g. $Secure:$SDS), the properties of the entry are set correctly and lstat() will work.

(DIS-1911)