Skip collecting /proc/self/fd/* symlinks on live systems
One of the fds in /proc/self/fd is a symlink to the tar file in which
all artefacts are collected. Collecting the file this symlink links to
will result in a "self" collect, which is not desired.

On certain systems a devtmpfs is also present on some other part of the
system (e.g. somewhere below /var) then /dev. One of the fd symlinks in
/dev/fd points to /proc/self/fd which will result in such a situation.
Note that collecting /proc with --proc does not suffer this issue as it
will not collect the file a symlink points to by design.

When a symlink is collected and its parent resolves to the same
directory /proc/self/fd resolves to (the self part is also a symlink
pointing to the pid of the acquire process), only the symlink will be
collected and not the location it points to.

(DIS-1785)